+++ app/src/main/java/ru/edgarakert/biblenote/data/db/AppDatabase.kt
@@ -1,14 +1,16 @@
 package ru.edgarakert.biblenote.data.db
 
 import android.content.Context
+import androidx.room.AutoMigration
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
 
 @Database(
     entities = [Note::class, Folder::class],
-    version = 1,
-    exportSchema = true
+    version = 2,
+    exportSchema = true,
+    autoMigrations = [AutoMigration(from = 1, to = 2)]
 )
 abstract class AppDatabase : RoomDatabase() {
     abstract fun noteDao(): NoteDao